requirements.txt
Django
psycopg2
djangorestframework-simplejwt[crypto]
djangorestframework
drf-spectacular